refactor(dashboard): split Preact client into components
