+++ src/features/dashboard/client/lib/preferences.js
@@ -1,0 +1,17 @@
+const PREFS_KEY = 'ldev.dashboard.prefs';
+
+export function readPrefs() {
+  try {
+    return JSON.parse(localStorage.getItem(PREFS_KEY) || '{}');
+  } catch {
+    return {};
+  }
+}
+
+export function writePrefs(prefs) {
+  try {
+    localStorage.setItem(PREFS_KEY, JSON.stringify(prefs));
+  } catch {
+    // Ignore private browsing and locked storage.
+  }
+}
