+++ tests/unit/dashboard-html.test.ts
@@ -4,7 +4,22 @@ import {describe, expect, test} from 'vitest';
 
 import {dashboardHtml} from '../../src/features/dashboard/dashboard-html.js';
 
-const dashboardClientApp = fs.readFileSync('src/features/dashboard/client/app.jsx', 'utf8');
+const dashboardClientFiles = [
+  'src/features/dashboard/client/app.jsx',
+  'src/features/dashboard/client/components/activity.jsx',
+  'src/features/dashboard/client/components/create-modal.jsx',
+  'src/features/dashboard/client/components/header.jsx',
+  'src/features/dashboard/client/components/maintenance.jsx',
+  'src/features/dashboard/client/components/modal.jsx',
+  'src/features/dashboard/client/components/simple-form-modal.jsx',
+  'src/features/dashboard/client/components/toolbar.jsx',
+  'src/features/dashboard/client/components/worktree-card.jsx',
+  'src/features/dashboard/client/components/worktree-sections.jsx',
+  'src/features/dashboard/client/lib/dashboard-state.js',
+  'src/features/dashboard/client/lib/preferences.js',
+  'src/features/dashboard/client/lib/tasks.js',
+];
+const dashboardClientSource = dashboardClientFiles.map((file) => fs.readFileSync(file, 'utf8')).join('\n');
 const dashboardClientStyles = fs.readFileSync('src/features/dashboard/client/styles.css', 'utf8');
 
 describe('dashboardHtml', () => {
@@ -14,38 +29,40 @@ describe('dashboardHtml', () => {
   });
 
   test('mounts a Preact dashboard app with the live dashboard APIs', () => {
-    expect(dashboardClientApp).toContain("from 'preact'");
-    expect(dashboardClientApp).toContain('function App()');
-    expect(dashboardClientApp).toContain("render(<App />, document.getElementById('dashboard-root'))");
-    expect(dashboardClientApp).toContain("new EventSource('/api/tasks/stream')");
-    expect(dashboardClientApp).toContain("fetch('/api/status'");
+    expect(dashboardClientSource).toContain("from 'preact'");
+    expect(dashboardClientSource).toContain('function App()');
+    expect(dashboardClientSource).toContain("render(<App />, document.getElementById('dashboard-root'))");
+    expect(dashboardClientSource).toContain("new EventSource('/api/tasks/stream')");
+    expect(dashboardClientSource).toContain("fetch('/api/status'");
   });
 
-  test('keeps card sections and dirty worktree context visible', () => {
-    expect(dashboardClientApp).toContain('function buildSections(wt)');
-    expect(dashboardClientApp).toContain("label: 'Changes'");
-    expect(dashboardClientApp).toContain("label: 'Services'");
-    expect(dashboardClientApp).toContain("label: 'Commits'");
-    expect(dashboardClientApp).toContain('Workspace details');
-    expect(dashboardClientApp).toContain('Latest commit');
-    expect(dashboardClientApp).toContain('`${wt.changedFiles} pending`');
+  test('keeps card sections and service state visible', () => {
+    expect(dashboardClientSource).toContain('function buildSections(wt)');
+    expect(dashboardClientSource).toContain("label: 'Changes'");
+    expect(dashboardClientSource).toContain("label: 'Services'");
+    expect(dashboardClientSource).toContain("label: 'Commits'");
+    expect(dashboardClientSource).toContain('Workspace details');
+    expect(dashboardClientSource).toContain('Latest commit');
+    expect(dashboardClientSource).toContain('`${changedFiles} pending`');
+    expect(dashboardClientSource).toContain('service.service || service.name');
+    expect(dashboardClientSource).toContain('serviceStatusLabel(service)');
     expect(dashboardClientStyles).toContain('card-chip-red');
   });
 
   test('includes guided modals for worktrees, databases, diagnostics, and resource exports', () => {
-    expect(dashboardClientApp).toContain('function CreateModal');
-    expect(dashboardClientApp).toContain('function SimpleFormModal');
-    expect(dashboardClientApp).toContain('title="DB tools"');
-    expect(dashboardClientApp).toContain('title="Resource export"');
-    expect(dashboardClientApp).toContain('/resource/export');
-    expect(dashboardClientApp).toContain('/api/doctor');
-    expect(dashboardClientApp).toContain('/mcp/setup');
-    expect(dashboardClientApp).toContain('/deploy/status');
-    expect(dashboardClientApp).toContain('/deploy/cache-update');
+    expect(dashboardClientSource).toContain('function CreateModal');
+    expect(dashboardClientSource).toContain('function SimpleFormModal');
+    expect(dashboardClientSource).toContain('title="DB tools"');
+    expect(dashboardClientSource).toContain('title="Resource export"');
+    expect(dashboardClientSource).toContain('/resource/export');
+    expect(dashboardClientSource).toContain('/api/doctor');
+    expect(dashboardClientSource).toContain('/mcp/setup');
+    expect(dashboardClientSource).toContain('/deploy/status');
+    expect(dashboardClientSource).toContain('/deploy/cache-update');
   });
 
   test('does not show duplicate primary actions in running worktree cards', () => {
-    expect(dashboardClientApp).toContain("primary[0] !== 'start' && !running");
-    expect(dashboardClientApp).toContain('wt.env && !stopped');
+    expect(dashboardClientSource).toContain("primary[0] !== 'start' && !running");
+    expect(dashboardClientSource).toContain('wt.env && !stopped');
   });
 });
